getting nutrition plan by user

+++ CaloryCalculation.Application/DTOs/Goal/GoalCreationDto.cs
@@ -5,7 +5,7 @@ namespace CaloryCalculation.Application.DTOs.Goal;
 public class GoalCreationDto
 {
     public Gender Gender { get; set; }
-    public float Weight { get; set; }
+    public double Weight { get; set; }
     public GoalType Goal { get; set; }
     public ActivityLevel ActivityLevel { get; set; }
     public int? UserId { get; set; }
